Updates to splash screen
