+++ manifest-editor/src/apps/Splash/Splash.tsx
@@ -1,9 +1,8 @@
 import { FlexContainerColumn } from "../../components/layout/FlexContainer";
 import { useShell } from "../../context/ShellContext/ShellContext";
 import { RecentFiles } from "../../components/widgets/RecentFiles";
-import { NewTemplates } from "../../components/widgets/NewTemplates";
 import { PaddingComponentLarge } from "../../atoms/PaddingComponent";
-import { LoadManifest } from "../../components/widgets/LoadManifest";
+import { LoadManifest } from "../../components/widgets/LoadManifest/LoadManifest";
 import { Layout } from "../../shell/Layout/Layout";
 import { LayoutPanel } from "../../shell/Layout/Layout.types";
 import { ManifestEditorProvider } from "../ManifestEditor/ManifestEditor.context";
@@ -15,6 +14,7 @@ export function Splash() {
     shellContext.changeResourceID(id);
     shellContext.changeSelectedApplication("ManifestEditor");
   };
+
   const leftPanels: LayoutPanel[] = [];
 
   const centerPanels: LayoutPanel[] = [
@@ -23,14 +23,17 @@ export function Splash() {
       label: "Splash Screen",
       icon: "",
       render: () => (
-        <FlexContainerColumn>
-          <PaddingComponentLarge />
-          <LoadManifest />
-          <PaddingComponentLarge />
-          <NewTemplates newTemplates={shellContext.newTemplates} changeManifest={handleClick} />
-          <PaddingComponentLarge />
-          <RecentFiles recentManifests={shellContext.recentManifests} changeManifest={handleClick} />
-          <PaddingComponentLarge />
+        <FlexContainerColumn style={{ height: "100%", justifyContent: "space-between" }}>
+          <div>
+            <PaddingComponentLarge />
+            <LoadManifest />
+            <PaddingComponentLarge />
+          </div>
+          <div style={{ width: "100%", height: "50%", backgroundColor: "white" }}>
+            <FlexContainerColumn style={{ width: "90%", margin: "auto" }}>
+              <RecentFiles recentManifests={shellContext.recentManifests} changeManifest={handleClick} />
+            </FlexContainerColumn>
+          </div>
         </FlexContainerColumn>
       ),
     },
